use uint32_t instead of uint64_t; codal_target_hal.h

+++ inc/core/codal_target_hal.h
@@ -35,9 +35,9 @@ extern "C"
 
     void target_reset();
 
-    void target_wait(unsigned long milliseconds);
+    void target_wait(uint32_t milliseconds);
 
-    void target_wait_us(unsigned long us);
+    void target_wait_us(uint32_t us);
 
     int target_seed_random(uint32_t rand);
